Refactor Milkdrop overlay panels
